// src/app/models/product-model/product-content-model.ts
export class ProductContentModel {
    productName: string;
    manufacturerId: string;
    measureUnitId: string;
    typeId: number;
    qtyPerUnit: number;
    weight: number;
    groupAttributeId: number;
    canDeliverMultiCity: boolean;
    requireInstall: boolean;
    requireSerial: boolean;
    fullDescription: string;
    shortDescription: string;
}